Update paths and Create page
Separate into individual pages, use navigate, add individual paths
