+++ client/src/pages/create/Entity.tsx
@@ -2,76 +2,29 @@
 import React, { useEffect, useState } from "react";
 
 // Existing and custom components
-import {
-  Button,
-  Card,
-  CardBody,
-  CardFooter,
-  CardHeader,
-  Checkbox,
-  CheckboxGroup,
-  Flex,
-  FormControl,
-  FormErrorMessage,
-  FormHelperText,
-  FormLabel,
-  Heading,
-  Input,
-  List,
-  ListItem,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalHeader,
-  ModalOverlay,
-  Select,
-  Stack,
-  StackDivider,
-  Tag,
-  TagCloseButton,
-  Text,
-  Textarea,
-  Tooltip,
-  useDisclosure,
-  useToast,
-} from "@chakra-ui/react";
+import { Button, Checkbox, CheckboxGroup, Flex, FormControl, FormErrorMessage, FormHelperText, FormLabel, Heading, Input, Modal, ModalBody, ModalCloseButton, ModalContent, ModalHeader, ModalOverlay, Select, Stack, Tag, TagCloseButton, Text, Textarea, useDisclosure, useToast } from "@chakra-ui/react";
 import { Content } from "@components/Container";
-import Attribute from "@components/Attribute";
 import Error from "@components/Error";
 import Icon from "@components/Icon";
 import Linky from "@components/Linky";
+import Attribute from "@components/Attribute";
 import Loading from "@components/Loading";
-import Values from "@components/Values";
 
 // Existing and custom types
-import {
-  IAttribute,
-  AttributeModel,
-  AttributeProps,
-  CollectionModel,
-  IEntity,
-  EntityModel,
-  IValue,
-  ICollection,
-  CreatePage,
-} from "@types";
+import { AttributeModel, AttributeProps, CollectionModel, EntityModel, IEntity } from "@types";
+
+// Utility functions and libraries
+import { getData, postData } from "@database/functions";
+import { useToken } from "src/authentication/useToken";
+import { isValidAttributes } from "src/util";
+import _ from "lodash";
+import dayjs from "dayjs";
+import { nanoid } from "nanoid";
 
 // Routing and navigation
 import { useNavigate } from "react-router-dom";
 
-// Utility functions and libraries
-import { isValidAttributes, isValidValues } from "src/functions";
-import { getData, postData } from "@database/functions";
-import _ from "lodash";
-import { nanoid } from "nanoid";
-import dayjs from "dayjs";
-import { useToken } from "src/authentication/useToken";
-
-const EntityPage = (props: {
-  createPageState: CreatePage;
-  setCreatePageState: React.Dispatch<React.SetStateAction<CreatePage>>;
-}) => {
+const Entity = () => {
   // Used to manage what detail inputs are presented
   const [pageState, setPageState] = useState(
     "start" as "start" | "attributes" | "associations"
@@ -246,11 +199,6 @@ const EntityPage = (props: {
     } else if (_.isEqual("attributes", pageState)) {
       setPageState("associations");
     }
-  };
-
-  // Handle clicking "Cancel"
-  const onCancel = () => {
-    props.setCreatePageState("default");
   };
 
   // Removal callback
@@ -712,7 +660,7 @@ const EntityPage = (props: {
                     colorScheme={"red"}
                     variant={"outline"}
                     rightIcon={<Icon name={"cross"} />}
-                    onClick={onCancel}
+                    onClick={() => navigate("/entities")}
                   >
                     Cancel
                   </Button>
@@ -768,724 +716,4 @@ const EntityPage = (props: {
   );
 };
 
-const CollectionPage = (props: {
-  createPageState: CreatePage;
-  setCreatePageState: React.Dispatch<React.SetStateAction<CreatePage>>;
-}) => {
-  const navigate = useNavigate();
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const [token, _useToken] = useToken();
-
-  const [type, setType] = useState("collection" as "collection" | "project");
-  const [name, setName] = useState("");
-  const [created, setCreated] = useState(
-    dayjs(Date.now()).format("YYYY-MM-DDTHH:mm")
-  );
-  const [owner, setOwner] = useState(token.username);
-  const [description, setDescription] = useState("");
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  // Form validation
-  const isNameError = name === "";
-  const isOwnerError = owner === "";
-  const isDescriptionError = description === "";
-  const isDetailsError = isNameError || isOwnerError || isDescriptionError;
-
-  const collectionData: ICollection = {
-    type: type,
-    name: name,
-    description: description,
-    owner: owner,
-    created: created,
-    collections: [],
-    entities: [],
-    history: [],
-  };
-
-  return (
-    <Content>
-      <Flex
-        direction={"column"}
-        justify={"center"}
-        p={"2"}
-        gap={"6"}
-        maxW={"7xl"}
-        wrap={"wrap"}
-      >
-        <Flex
-          direction={"column"}
-          w={"100%"}
-          p={"4"}
-          bg={"white"}
-          rounded={"md"}
-        >
-          {/* Page header */}
-          <Flex direction={"column"} p={"2"} pt={"4"} pb={"4"}>
-            <Flex direction={"row"} align={"center"} justify={"space-between"}>
-              <Heading fontWeight={"semibold"}>Create Collection</Heading>
-              <Button
-                rightIcon={<Icon name={"info"} />}
-                variant={"outline"}
-                onClick={onOpen}
-              >
-                Info
-              </Button>
-            </Flex>
-          </Flex>
-
-          {/* Data input */}
-          <Flex
-            direction={"row"}
-            justify={"center"}
-            gap={"6"}
-            p={"2"}
-            pb={"6"}
-            mb={["12", "8"]}
-          >
-            <Flex
-              direction={"column"}
-              gap={"2"}
-              w={"100%"}
-              maxW={"4xl"}
-            >
-              <Heading fontWeight={"semibold"} size={"lg"}>
-                Details
-              </Heading>
-              <Text>Specify some basic details about this Collection.</Text>
-              <Flex direction="row" gap={"4"} wrap={["wrap", "nowrap"]}>
-                <FormControl isRequired isInvalid={isNameError}>
-                  <FormLabel htmlFor="name" fontWeight={"normal"}>
-                    Name
-                  </FormLabel>
-                  <Input
-                    id="name"
-                    name="name"
-                    borderColor={"blackAlpha.300"}
-                    focusBorderColor={"black"}
-                    value={name}
-                    onChange={(event) => setName(event.target.value)}
-                  />
-                  {isNameError && (
-                    <FormErrorMessage>
-                      A name or ID must be specified.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
-
-                <FormControl isRequired isInvalid={isOwnerError}>
-                  <FormLabel htmlFor="owner" fontWeight={"normal"}>
-                    Owner
-                  </FormLabel>
-                  <Input
-                    id="owner"
-                    name="owner"
-                    borderColor={"blackAlpha.300"}
-                    focusBorderColor={"black"}
-                    value={owner}
-                    onChange={(event) => setOwner(event.target.value)}
-                  />
-                  {isOwnerError && (
-                    <FormErrorMessage>
-                      An owner of the Collection is required.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
-              </Flex>
-
-              <Flex direction="row" gap={"4"} wrap={["wrap", "nowrap"]}>
-                <FormControl>
-                  <FormLabel htmlFor="date" fontWeight={"normal"}>
-                    Created
-                  </FormLabel>
-
-                  <Input
-                    size={"md"}
-                    type={"datetime-local"}
-                    value={created}
-                    onChange={(event) =>
-                      setCreated(
-                        dayjs(event.target.value).format("YYYY-MM-DDTHH:mm")
-                      )
-                    }
-                  />
-                </FormControl>
-
-                <FormControl isRequired isInvalid={isDescriptionError}>
-                  <FormLabel htmlFor="description" fontWeight={"normal"}>
-                    Description
-                  </FormLabel>
-                  <Textarea
-                    id="description"
-                    name="description"
-                    value={description}
-                    onChange={(event) => setDescription(event.target.value)}
-                  />
-                  {isDescriptionError && (
-                    <FormErrorMessage>
-                      A description must be provided.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
-              </Flex>
-
-              <Flex direction="row" gap={"4"} wrap={["wrap", "nowrap"]}>
-                <FormControl>
-                  <FormLabel htmlFor="date" fontWeight={"normal"}>
-                    Collection Type
-                  </FormLabel>
-                  <CheckboxGroup>
-                    <Flex direction={"column"} gap={"4"}>
-                      <Checkbox
-                        isChecked={_.isEqual(type, "collection")}
-                        onChange={(_event) => {
-                          setType("collection");
-                        }}
-                      >
-                        <Tooltip
-                          label={
-                            "A standard Collection with no special properties."
-                          }
-                        >
-                          <Flex direction={"row"} gap={"4"} align={"center"}>
-                            Standard
-                            <Icon name={"info"} />
-                          </Flex>
-                        </Tooltip>
-                      </Checkbox>
-                      <Checkbox
-                        isChecked={_.isEqual(type, "project")}
-                        onChange={(_event) => {
-                          setType("project");
-                        }}
-                      >
-                        <Tooltip label={"A Collection denoted as a Project."}>
-                          <Flex direction={"row"} gap={"4"} align={"center"}>
-                            Project
-                            <Icon name={"info"} />
-                          </Flex>
-                        </Tooltip>
-                      </Checkbox>
-                    </Flex>
-                  </CheckboxGroup>
-                </FormControl>
-              </Flex>
-            </Flex>
-          </Flex>
-
-          {/* Action buttons */}
-          <Flex
-            direction={"row"}
-            flexWrap={"wrap"}
-            gap={"6"}
-            justify={"space-between"}
-            alignSelf={"center"}
-            w={"100%"}
-            p={"4"}
-          >
-            <Button
-              colorScheme={"red"}
-              rightIcon={<Icon name={"cross"} />}
-              variant={"outline"}
-              onClick={() => props.setCreatePageState("default")}
-            >
-              Cancel
-            </Button>
-
-            <Button
-              colorScheme={"green"}
-              rightIcon={<Icon name={"check"} />}
-              onClick={() => {
-                // Push the data
-                setIsSubmitting(true);
-                postData(`/collections/create`, collectionData).then(() => {
-                  setIsSubmitting(false);
-                  navigate("/collections");
-                });
-              }}
-              isDisabled={isDetailsError && !isSubmitting}
-            >
-              Finish
-            </Button>
-          </Flex>
-        </Flex>
-      </Flex>
-
-      {/* Information modal */}
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>Collections</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            <Flex direction={"column"} gap={"4"} p={"2"}>
-              <Text>
-                Collections can be used to organize Entities. Any type of Entity
-                can be included in a Collection. Entities can be added and
-                removed from a Collection after it has been created.
-              </Text>
-            </Flex>
-          </ModalBody>
-        </ModalContent>
-      </Modal>
-    </Content>
-  );
-};
-
-const AttributePage = (props: {
-  createPageState: CreatePage;
-  setCreatePageState: React.Dispatch<React.SetStateAction<CreatePage>>;
-}) => {
-  const navigate = useNavigate();
-  const { isOpen, onOpen, onClose } = useDisclosure();
-
-  const [name, setName] = useState("");
-  const [description, setDescription] = useState("");
-  const [values, setValues] = useState([] as IValue<any>[]);
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  // Various validation error states
-  const isNameError = name === "";
-  const isDescriptionError = description === "";
-  const [isValueError, setIsValueError] = useState(false);
-  const isDetailsError = isNameError || isDescriptionError || isValueError;
-
-  const attributeData: IAttribute = {
-    name: name,
-    description: description,
-    values: values,
-  };
-
-  // Check the Values for errors each time they update
-  useEffect(() => {
-    setIsValueError(!isValidValues(values, true) || values.length === 0);
-  }, [values]);
-
-  const onSubmit = () => {
-    setIsSubmitting(true);
-
-    // Push the data
-    postData(`/attributes/create`, attributeData).then(() => {
-      setIsSubmitting(false);
-      navigate("/attributes");
-    });
-  };
-
-  return (
-    <Content>
-      <Flex direction={"column"} p={"2"} gap={"6"} wrap={"wrap"}>
-        <Flex direction={"column"} w={"100%"} p={"4"} bg={"white"}>
-          {/* Page header */}
-          <Flex direction={"column"} p={"2"} pt={"4"} pb={"4"}>
-            <Flex direction={"row"} align={"center"} justify={"space-between"}>
-              <Heading fontWeight={"semibold"}>Create Attribute</Heading>
-              <Button
-                rightIcon={<Icon name={"info"} />}
-                variant={"outline"}
-                onClick={onOpen}
-              >
-                Info
-              </Button>
-            </Flex>
-          </Flex>
-
-          <Flex
-            direction={"column"}
-            justify={"center"}
-            align={"center"}
-            gap={"6"}
-            p={"2"}
-            pb={"6"}
-            mb={["12", "8"]}
-          >
-            <Flex direction={"column"} w={"100%"}>
-              <Heading fontWeight={"semibold"} size={"lg"}>
-                Details
-              </Heading>
-              <Text>
-                Specify some basic details about this template Attribute. The
-                metadata associated with this template should be specified using
-                Values.
-              </Text>
-            </Flex>
-
-            <Flex
-              direction={"column"}
-              gap={"2"}
-              w={"100%"}
-              maxW={"4xl"}
-              wrap={["wrap", "nowrap"]}
-            >
-              <Flex direction={"row"} gap={"4"} wrap={["wrap", "nowrap"]}>
-                <FormControl isRequired>
-                  <FormLabel>Name</FormLabel>
-                  <Input
-                    placeholder={"Name"}
-                    value={name}
-                    onChange={(event) => setName(event.target.value)}
-                    required
-                  />
-                  {isNameError && (
-                    <FormErrorMessage>
-                      A name must be specified for the Attribute.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
-
-                <FormControl isRequired>
-                  <FormLabel>Description</FormLabel>
-                  <Textarea
-                    value={description}
-                    placeholder={"Attribute Description"}
-                    onChange={(event) => setDescription(event.target.value)}
-                  />
-                  {isDescriptionError && (
-                    <FormErrorMessage>
-                      A description should be provided for the Attribute.
-                    </FormErrorMessage>
-                  )}
-                </FormControl>
-              </Flex>
-
-              <Flex>
-                <FormControl isRequired isInvalid={isValueError}>
-                  <FormLabel>Values</FormLabel>
-                  <Values
-                    collection={values}
-                    viewOnly={false}
-                    setValues={setValues}
-                  />
-                </FormControl>
-              </Flex>
-            </Flex>
-          </Flex>
-
-          {/* Action buttons */}
-          <Flex
-            direction={"row"}
-            flexWrap={"wrap"}
-            gap={"6"}
-            justify={"space-between"}
-            alignSelf={"center"}
-            w={"100%"}
-            p={"4"}
-          >
-            <Button
-              colorScheme={"red"}
-              variant={"outline"}
-              rightIcon={<Icon name={"cross"} />}
-              onClick={() => props.setCreatePageState("default")}
-            >
-              Cancel
-            </Button>
-            <Button
-              colorScheme={"green"}
-              rightIcon={<Icon name={"check"} />}
-              onClick={onSubmit}
-              isDisabled={isDetailsError && !isSubmitting}
-            >
-              Finish
-            </Button>
-          </Flex>
-        </Flex>
-      </Flex>
-
-      {/* Information modal */}
-      <Modal isOpen={isOpen} onClose={onClose}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>Attributes</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            <Flex direction={"column"} gap={"4"} p={"2"}>
-              <Text>There are six supported types of metadata:</Text>
-              <List spacing={2}>
-                <ListItem>
-                  <Flex gap={"2"} align={"center"}>
-                    <Icon name={"v_date"} color={"orange.300"} />
-                    <Text>Date: Used to specify a point in time.</Text>
-                  </Flex>
-                </ListItem>
-                <ListItem>
-                  <Flex gap={"2"} align={"center"}>
-                    <Icon name={"v_text"} color={"blue.300"} />
-                    <Text>Text: Used to specify text of variable length.</Text>
-                  </Flex>
-                </ListItem>
-                <ListItem>
-                  <Flex gap={"2"} align={"center"}>
-                    <Icon name={"v_number"} color={"green.300"} />
-                    <Text>Number: Used to specify a numerical value.</Text>
-                  </Flex>
-                </ListItem>
-                <ListItem>
-                  <Flex gap={"2"} align={"center"}>
-                    <Icon name={"v_url"} color={"yellow.300"} />
-                    <Text>URL: Used to specify a link.</Text>
-                  </Flex>
-                </ListItem>
-                <ListItem>
-                  <Flex gap={"2"} align={"center"}>
-                    <Icon name={"entity"} color={"purple.300"} />
-                    <Text>
-                      Entity: Used to specify a relation to another Entity.
-                    </Text>
-                  </Flex>
-                </ListItem>
-                <ListItem>
-                  <Flex gap={"2"} align={"center"}>
-                    <Icon name={"v_select"} color={"teal.300"} />
-                    <Text>
-                      Select: Used to specify an option from a group of options.
-                    </Text>
-                  </Flex>
-                </ListItem>
-              </List>
-              <Text>Values can be specified using the buttons.</Text>
-            </Flex>
-          </ModalBody>
-        </ModalContent>
-      </Modal>
-    </Content>
-  );
-};
-
-const Create = () => {
-  const [createPage, setCreatePage] = useState("default" as CreatePage);
-
-  return (
-    <>
-      {/* Default landing page for creating metadata */}
-      {_.isEqual(createPage, "default") && (
-        <Content>
-          <Flex direction={"column"} justify={"center"} gap={"6"} wrap={"wrap"}>
-            <Flex
-              direction={"column"}
-              w={"100%"}
-              p={"2"}
-              bg={"white"}
-              rounded={"md"}
-            >
-              <Flex
-                direction={"row"}
-                justify={"center"}
-                align={"center"}
-                gap={"6"}
-                p={"2"}
-                wrap={"wrap"}
-              >
-                {/* Collection card */}
-                <Card maxW={"sm"} h={"lg"} variant={"outline"}>
-                  <CardHeader>
-                    <Flex
-                      gap={"4"}
-                      w={"100%"}
-                      justify={"center"}
-                      align={"center"}
-                    >
-                      <Icon name={"collection"} size={"lg"} />
-                      <Heading size={"lg"}>Collection</Heading>
-                    </Flex>
-                  </CardHeader>
-                  <CardBody>
-                    <Stack divider={<StackDivider />} spacing={"2"}>
-                      <Flex
-                        p={"2"}
-                        gap={"4"}
-                        align={"center"}
-                        direction={"column"}
-                      >
-                        <Heading size={"xs"} textTransform={"uppercase"}>
-                          Description
-                        </Heading>
-                        <Text>
-                          Create a Collection to group and organize Entities.
-                        </Text>
-                      </Flex>
-                      <Flex
-                        p={"2"}
-                        gap={"4"}
-                        align={"center"}
-                        direction={"column"}
-                      >
-                        <Heading size={"xs"} textTransform={"uppercase"}>
-                          Fields
-                        </Heading>
-                        <Flex gap={"2"}>
-                          <Tag colorScheme={"red"}>Name</Tag>
-                          <Tag colorScheme={"red"}>Owner</Tag>
-                          <Tag colorScheme={"red"}>Description</Tag>
-                        </Flex>
-                      </Flex>
-                    </Stack>
-                  </CardBody>
-                  <CardFooter>
-                    <Flex w={"100%"} justify={"center"}>
-                      <Button
-                        colorScheme={"green"}
-                        rightIcon={<Icon name={"add"} />}
-                        onClick={() => setCreatePage("collection")}
-                      >
-                        Create
-                      </Button>
-                    </Flex>
-                  </CardFooter>
-                </Card>
-
-                {/* Entity card */}
-                <Card maxW={"sm"} h={"lg"} variant={"outline"}>
-                  <CardHeader>
-                    <Flex
-                      gap={"4"}
-                      w={"100%"}
-                      justify={"center"}
-                      align={"center"}
-                    >
-                      <Icon name={"entity"} size={"lg"} />
-                      <Heading size={"lg"}>Entity</Heading>
-                    </Flex>
-                  </CardHeader>
-                  <CardBody>
-                    <Stack divider={<StackDivider />} spacing={"2"}>
-                      <Flex
-                        p={"2"}
-                        gap={"4"}
-                        align={"center"}
-                        direction={"column"}
-                      >
-                        <Heading size={"xs"} textTransform={"uppercase"}>
-                          Description
-                        </Heading>
-                        <Text>
-                          Create an Entity to group metadata about a physical or
-                          digital resource.
-                        </Text>
-                      </Flex>
-                      <Flex
-                        p={"2"}
-                        gap={"4"}
-                        align={"center"}
-                        direction={"column"}
-                      >
-                        <Heading size={"xs"} textTransform={"uppercase"}>
-                          Fields
-                        </Heading>
-                        <Flex gap={"2"} wrap={"wrap"}>
-                          <Tag colorScheme={"red"}>Name</Tag>
-                          <Tag colorScheme={"red"}>Owner</Tag>
-                          <Tag colorScheme={"red"}>Created</Tag>
-                          <Tag colorScheme={"teal"}>Description</Tag>
-                          <Tag colorScheme={"teal"}>Origins</Tag>
-                          <Tag colorScheme={"teal"}>Products</Tag>
-                          <Tag colorScheme={"teal"}>Collections</Tag>
-                          <Tag colorScheme={"teal"}>Attributes</Tag>
-                        </Flex>
-                      </Flex>
-                    </Stack>
-                  </CardBody>
-                  <CardFooter>
-                    <Flex w={"100%"} justify={"center"}>
-                      <Button
-                        colorScheme={"green"}
-                        rightIcon={<Icon name={"add"} />}
-                        onClick={() => setCreatePage("entity")}
-                      >
-                        Create
-                      </Button>
-                    </Flex>
-                  </CardFooter>
-                </Card>
-
-                {/* Attribute card */}
-                <Card maxW={"sm"} h={"lg"} variant={"outline"}>
-                  <CardHeader>
-                    <Flex
-                      gap={"4"}
-                      w={"100%"}
-                      justify={"center"}
-                      align={"center"}
-                    >
-                      <Icon name={"attribute"} size={"lg"} />
-                      <Heading size={"lg"}>Template Attribute</Heading>
-                    </Flex>
-                  </CardHeader>
-                  <CardBody>
-                    <Stack divider={<StackDivider />} spacing={"2"}>
-                      <Flex
-                        p={"2"}
-                        gap={"4"}
-                        align={"center"}
-                        direction={"column"}
-                      >
-                        <Heading size={"xs"} textTransform={"uppercase"}>
-                          Description
-                        </Heading>
-                        <Text>
-                          Create a template Attribute to standardize reusable
-                          components of metadata to be associated with Entities.
-                        </Text>
-                      </Flex>
-                      <Flex
-                        p={"2"}
-                        gap={"4"}
-                        align={"center"}
-                        direction={"column"}
-                      >
-                        <Heading size={"xs"} textTransform={"uppercase"}>
-                          Fields
-                        </Heading>
-                        <Flex gap={"2"} wrap={"wrap"}>
-                          <Tag colorScheme={"red"}>Name</Tag>
-                          <Tag colorScheme={"red"}>Description</Tag>
-                          <Tag colorScheme={"red"}>Values</Tag>
-                        </Flex>
-                      </Flex>
-                    </Stack>
-                  </CardBody>
-                  <CardFooter>
-                    <Flex w={"100%"} justify={"center"}>
-                      <Button
-                        colorScheme={"green"}
-                        rightIcon={<Icon name={"add"} />}
-                        onClick={() => setCreatePage("attribute")}
-                      >
-                        Create
-                      </Button>
-                    </Flex>
-                  </CardFooter>
-                </Card>
-              </Flex>
-            </Flex>
-          </Flex>
-        </Content>
-      )}
-
-      {/* Create an Entity */}
-      {_.isEqual(createPage, "entity") && (
-        <EntityPage
-          createPageState={createPage}
-          setCreatePageState={setCreatePage}
-        />
-      )}
-
-      {/* Create a Collection */}
-      {_.isEqual(createPage, "collection") && (
-        <CollectionPage
-          createPageState={createPage}
-          setCreatePageState={setCreatePage}
-        />
-      )}
-
-      {/* Create an Attribute */}
-      {_.isEqual(createPage, "attribute") && (
-        <AttributePage
-          createPageState={createPage}
-          setCreatePageState={setCreatePage}
-        />
-      )}
-    </>
-  );
-};
-
-export default Create;
+export default Entity;
