+++ client/src/pages/view/Attribute.tsx
