+++ client/src/pages/view/Collection.tsx
